requirements.txt
# ============================================================
# RAG PDF Assistant — Requirements
# ============================================================

# Core LangChain + integrations
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.7
langchain-huggingface==0.1.2
langchain-ollama==0.2.1
langchain-text-splitters==0.3.2

# PDF processing
pypdf==4.3.1
pdfplumber==0.11.4

# Embeddings & vector store
sentence-transformers==3.2.1
faiss-cpu==1.9.0
torch==2.4.1
transformers==4.45.2

# Streamlit UI
streamlit==1.39.0
streamlit-extras==0.4.7

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
numpy==1.26.4
pandas==2.2.3

# Caching & performance
diskcache==5.6.3
